# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'@project.name@'","description":"'@project.description@'","contact":{"name":"Joe Doe","email":"dumy@dumy"},"version":"'@project.version@'"},"servers":[{"url":"http://localhost:8080","description":"'@project.description@'"}],"paths":{"/expenses":{"put":{"tags":["expense-controller"],"summary":"Edit a expense","description":"Edit a expense with the given details and optional file","operationId":"putExpense","requestBody":{"content":{"multipart/form-data":{"schema":{"required":["expense","userId"],"type":"object","properties":{"expense":{"$ref":"#/components/schemas/DtoRequestExpense"},"file":{"type":"string","format":"binary"},"userId":{"type":"integer","format":"int32"}}}}}},"responses":{"200":{"description":"Expense updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DtoResponseExpense"}}}}}},"post":{"tags":["expense-controller"],"summary":"Create a new expense","description":"Creates a new expense with the given details and optional file","operationId":"postExpense","requestBody":{"content":{"multipart/form-data":{"schema":{"required":["expense","userId"],"type":"object","properties":{"expense":{"$ref":"#/components/schemas/DtoRequestExpense"},"file":{"type":"string","format":"binary"},"userId":{"type":"integer","format":"int32"}}}}}},"responses":{"200":{"description":"Expense created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DtoResponseExpense"}}}}}},"delete":{"tags":["expense-controller"],"summary":"Delete logic or expense","operationId":"deleteExpenseByIdLogicOrThrowException","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"Expense delete logic successfully","content":{"application/json":{}}}}}},"/categories/putById":{"put":{"tags":["expense-category-controller"],"operationId":"putExpenseCategory","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"description","in":"query","required":false,"schema":{"type":"string"}},{"name":"enabled","in":"query","required":false,"schema":{"type":"boolean"}},{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DtoCategory"}}}}}}},"/categories/postCategory":{"post":{"tags":["expense-category-controller"],"operationId":"postCategory","parameters":[{"name":"description","in":"query","required":true,"schema":{"type":"string"}},{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DtoCategory"}}}}}}},"/billexpenses/generate":{"post":{"tags":["bill-expense-controller"],"summary":"Generate bill expenses for a given period","description":"Generates bill expenses for the provided period. Validates that the period is in the past and that no overlapping bill already exists.","operationId":"generateExpenses","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PeriodDto"}}},"required":true},"responses":{"200":{"description":"Bill expenses generated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BillExpenseDto"}}}},"409":{"description":"Conflict - Overlapping period detected","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal Server Error - Unexpected server error or failed to process the bill expenses","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"400":{"description":"Bad Request - Invalid period or validation failure","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/reportchart/yearmonth":{"get":{"tags":["expense-report-chart-controller"],"operationId":"getYearMonth","parameters":[{"name":"start_year","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"end_year","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExpenseYearDto"}}}}}}}},"/reportchart/lastBillRecord":{"get":{"tags":["expense-report-chart-controller"],"operationId":"getLastBillRecord","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LastBillRecordDto"}}}}}}},"/reportchart/expenseByTypeAndCategory":{"get":{"tags":["expense-report-chart-controller"],"operationId":"get","parameters":[{"name":"start_date","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"end_date","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DtoExpenseKPI"}}}}}}}},"/reportchart/categoriesperiod":{"get":{"tags":["expense-report-chart-controller"],"operationId":"getCategoriesPeriod","parameters":[{"name":"start_date","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"end_date","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExpenseCategoryPeriodDto"}}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/expenses/getById":{"get":{"tags":["expense-controller"],"summary":"Get expense by id","operationId":"getExpenseById","parameters":[{"name":"expenseId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"Expenses retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DtoExpenseQuery"}}}}}}},"/expenses/getByFilters":{"get":{"tags":["expense-controller"],"summary":"Get expenses by filters","operationId":"getExpenses","parameters":[{"name":"dateFrom","in":"query","required":true,"schema":{"type":"string"}},{"name":"dateTo","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Expenses retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DtoExpenseQuery"}}}}}}},"/categories/getById/{id}":{"get":{"tags":["expense-category-controller"],"operationId":"getExpenseCategory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DtoCategory"}}}}}}},"/categories/all":{"get":{"tags":["expense-category-controller"],"operationId":"getAllCategories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DtoCategory"}}}}}}}},"/api/expenses/distributions/getAllByOwnerId":{"get":{"tags":["expense-distribution-controller"],"summary":"Get ALL Expenses by Id","description":"Get all expenses by Id","operationId":"getAllExpensesById","parameters":[{"name":"id","in":"query","description":"ID of the User","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"startDate","in":"query","description":"Start date in format yyyy-MM-dd","required":true,"schema":{"type":"string"}},{"name":"endDate","in":"query","description":"End date in format yyyy-MM-dd","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"get all successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExpenseOwnerVisualizerDTO"}}}}}}},"/expenses/note_credit":{"delete":{"tags":["expense-controller"],"summary":"Create note of credit","operationId":"createNoteOfCredit","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"204":{"description":"Note of credit created successfully","content":{"application/json":{}}}}}}},"components":{"schemas":{"DtoDistribution":{"type":"object","properties":{"ownerId":{"type":"integer","format":"int32"},"proportion":{"type":"number"}}},"DtoRequestExpense":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"},"providerId":{"type":"integer","format":"int32"},"expenseDate":{"type":"string","format":"date"},"invoiceNumber":{"type":"string"},"typeExpense":{"type":"string"},"categoryId":{"type":"integer","format":"int32"},"amount":{"type":"number"},"installments":{"type":"integer","format":"int32"},"distributions":{"type":"array","items":{"$ref":"#/components/schemas/DtoDistribution"}}}},"DtoCategory":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"},"lastUpdatedDatetime":{"type":"string","format":"date-time"},"state":{"type":"string"}}},"DtoInstallment":{"type":"object","properties":{"installmentNumber":{"type":"integer","format":"int32"},"paymentDate":{"type":"string","format":"date"}}},"DtoResponseExpense":{"type":"object","properties":{"description":{"type":"string"},"providerId":{"type":"integer","format":"int32"},"expenseDate":{"type":"string","format":"date"},"fileId":{"type":"string","format":"uuid"},"invoiceNumber":{"type":"string"},"expenseType":{"type":"string","enum":["COMUN","INDIVIDUAL","EXTRAORDINARIO","NOTE_OF_CREDIT"]},"dtoCategory":{"$ref":"#/components/schemas/DtoCategory"},"dtoDistributionList":{"type":"array","items":{"$ref":"#/components/schemas/DtoDistribution"}},"dtoInstallmentList":{"type":"array","items":{"$ref":"#/components/schemas/DtoInstallment"}}},"description":"Expense Response Data"},"PeriodDto":{"type":"object","properties":{"start_date":{"type":"string","format":"date"},"end_date":{"type":"string","format":"date"}}},"BillExpenseDto":{"type":"object","properties":{"start_date":{"type":"string","format":"date"},"end_date":{"type":"string","format":"date"},"bill_expense_id":{"type":"integer","format":"int32"},"owners":{"type":"array","items":{"$ref":"#/components/schemas/BillOwnerDto"}}}},"BillOwnerDto":{"type":"object","properties":{"owner_id":{"type":"integer","format":"int32"},"field_size":{"type":"integer","format":"int32"},"fines":{"type":"array","items":{"$ref":"#/components/schemas/FineDto"}},"expenses_common":{"type":"array","items":{"$ref":"#/components/schemas/ItemDto"}},"expenses_extraordinary":{"type":"array","items":{"$ref":"#/components/schemas/ItemDto"}},"expenses_individual":{"type":"array","items":{"$ref":"#/components/schemas/ItemDto"}},"notes_credit":{"type":"array","items":{"$ref":"#/components/schemas/ItemDto"}}}},"FineDto":{"type":"object","properties":{"fine_id":{"type":"integer","format":"int32"},"plot_id":{"type":"integer","format":"int32"},"description":{"type":"string"},"amount":{"type":"number"}}},"ItemDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"},"amount":{"type":"number"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"ExpenseYearDto":{"type":"object","properties":{"year":{"type":"integer","format":"int32"},"month":{"type":"integer","format":"int32"},"amount":{"type":"number"},"expense_type":{"type":"string"},"providerId":{"type":"integer","format":"int32"},"categoryId":{"type":"integer","format":"int32"}}},"DtoExpenseKPI":{"type":"object","properties":{"expenseType":{"type":"string"},"categoryId":{"type":"integer","format":"int32"},"description":{"type":"string"},"amount":{"type":"number"},"providerId":{"type":"integer","format":"int32"}}},"LastBillRecordDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"bills":{"type":"array","items":{"$ref":"#/components/schemas/DtoExpenseKPI"}},"fineAmount":{"type":"number"},"pendingAmount":{"type":"number"},"start_date":{"type":"string","format":"date"},"end_date":{"type":"string","format":"date"}}},"ExpenseCategoryPeriodDto":{"type":"object","properties":{"category":{"type":"string"},"amount":{"type":"number"}}},"DtoExpenseDistributionQuery":{"type":"object","properties":{"ownerId":{"type":"integer","format":"int32"},"ownerFullName":{"type":"string"},"amount":{"type":"number"},"proportion":{"type":"number"}}},"DtoExpenseInstallment":{"type":"object","properties":{"paymentDate":{"type":"string","format":"date"},"installmentNumber":{"type":"integer","format":"int32"}}},"DtoExpenseQuery":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"},"category":{"type":"string"},"categoryId":{"type":"integer","format":"int32"},"provider":{"type":"string"},"providerId":{"type":"integer","format":"int32"},"amount":{"type":"number"},"expenseType":{"type":"string"},"expenseDate":{"type":"string","format":"date"},"fileId":{"type":"string"},"distributionList":{"type":"array","items":{"$ref":"#/components/schemas/DtoExpenseDistributionQuery"}},"installmentList":{"type":"array","items":{"$ref":"#/components/schemas/DtoExpenseInstallment"}},"invoiceNumber":{"type":"string"}}},"ExpenseCategoryDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"}}},"ExpenseOwnerVisualizerDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"expenseId":{"type":"integer","format":"int32"},"description":{"type":"string"},"providerId":{"type":"integer","format":"int32"},"providerDescription":{"type":"string"},"expenseDate":{"type":"string","format":"date"},"fileId":{"type":"string","format":"uuid"},"invoiceNumber":{"type":"string"},"expenseType":{"type":"string","enum":["COMUN","INDIVIDUAL","EXTRAORDINARIO","NOTE_OF_CREDIT"]},"category":{"$ref":"#/components/schemas/ExpenseCategoryDTO"},"amount":{"type":"number"},"proportion":{"type":"number"},"installments":{"type":"integer","format":"int32"},"enabled":{"type":"boolean"}}}}}}